Implemented incremental learning on forest
Removed magic numbers in forest.py and tree.py. These should now be
based soley off dataset information.
Storage of the predicted class probability in the nodes is now stored in
dictionaries. This allows the class predictions of different nodes to be
compared to each other, even if they only predict off the initial
training data.